Add drawing context to support configurable y-axis flip, fix arc angle direction
